testing

+++ Heron.MudCalendar/Components/ItemPosition.cs
@@ -8,9 +8,9 @@ public class ItemPosition<[DynamicallyAccessedMembers(DynamicallyAccessedMemberT
     public int Position { get; set; }
     public int Total { get; set; }
     public DateOnly Date { get; set; }
-    public int Top { get; set; }
-    public int Left { get; set; }
-    public int Height { get; set; }
-    public int Width { get; set; } = 1;
-    public int Bottom => Top + Height;
+    public double Top { get; set; }
+    public double Left { get; set; }
+    public double Height { get; set; }
+    public double Width { get; set; } = 1;
+    public double Bottom => Top + Height;
 }
